Update serialization model for aggregation response. (#103)
* Add IVisitable to AggregationDictionary

* Add Visitor for AggregationDictionary

* Rename AggregationVisitor to AggregationContainerVisitor

* Move RangeAggregationExpressionVisitor

* Rename (and clean) TermsVisitor to TermsAggregationVisitor

* Rename RangeVisitor to RangeAggregationVisitor

* Rename (and clean) DateHistogramVisitor to DateHistogramAggregationVisitor

* Add AggregationDictionary to IVisitor interface

* Move MetricAggregationVisitor

* Remove Aggregations and AggregationsConverter

* Update int to long

* Delete Bucket and BucketsCollection

* Add IAggregate interface

* Add AggregateDictionary class

* Add ValueAggregate class

* Add BucketAggregate class

* Add PercentileAggregate classes

* Add IBucket interface

* Add BucketBase class

* Add KeyedBucket class

* Update DateHistogramBucket

* Update RangeBucket

* Add TermsBucket class

* Add AggregateFactory class

* Update BucketFactory class

* Update Aggregations to AggregateDictionary

* Update PercentileAggregate and PercentileAggregateConverter

* Remove unused using

* Remove BucketAggsJsonConverter

* Add TermsBucketConverter

* Update Aggregations Property.

* Update QueryData to get both primary aggregation key and type.

* Add default initialization to Buckets property.

* Add DateHistogramBucketConverter

* Update TermsBucketConverter

* Remove SummarizeByColumn

* Move hitsTotal query to main Visitor.

* Update docstrings and remove unused Json attributes.

* Add support for Keyed serialization.

* Remove dead code.

* Replace SummarizeByColumn by primaryKey and safety check on NaN values.

* Replace SummarizeByColumn by primaryKey and add CreateBucket delegate.

* Add JsonConverter for RangeBucket.

* Update RangeBucket Json Converter.

* Refactor response parser to support new serialization model.

* Update unit tests

* Fix SA1107: Code should not contain multiple statements on one line

* Update DateHistogramBucket Key from double to long.

* Update converters (check string is not null)

* Add new tests (code coverage).

* Add EncodeKustoField on TermsAggregationVistor

* Remove unused using

* Create a constant for %

* Add support for value_as_string (in case of timestamp usage)

* Update PercentileAggregateConverter to support value_as_string

* Add support of timestamp in ValueAggregate and PercentileAggregate

* Remove usage of generics in BucketAggregate

* Add TermsAggregate

* Use explicit cast instead as operator

* Refactoring bucket aggregation parsing.

* Update Unit Tests

* Update file names to be aligned with converter names.

* Add unit test  ParseElasticResponse_WithTermsAggs_ReturnsElasticResponseWithAggs

* Merge manualy changes on TermsAggregationVisitor

* Add additional e2e tests for metrics with no bucket aggregation

* Add ValueAggregateConverterTests unit test

* Fix typo issue

* Add PercentileAggregateConverterTests unit test

* Update trace logging

* Adding MetadataSeparator constant

* Update log trace

* Remove delegate

+++ K2Bridge/Models/Response/HitsCollection.cs
@@ -47,8 +47,8 @@ namespace K2Bridge.Models.Response
         /// <summary>
         /// Calculating HitsCollection.Total by summing up the DocCount values of every DateHistogramBucket.
         /// </summary>
-        /// <param name="docCount">int representing doc count.</param>
-        public void AddToTotal(int docCount)
+        /// <param name="docCount">long representing doc count.</param>
+        public void AddToTotal(long docCount)
         {
             Total.Value += docCount;
         }
@@ -56,7 +56,7 @@ namespace K2Bridge.Models.Response
         /// <summary>
         /// Set the HitsCollection.Total.
         /// </summary>
-        /// <param name="docCount">int representing doc count.</param>
+        /// <param name="docCount">long representing doc count.</param>
         public void SetTotal(long docCount)
         {
             Total.Value = docCount;
